Extract external config helpers into spack.externals_config
Move create_external_parser() from spack.solver.reuse and
external_config_with_implicit_externals() from spack.solver.runtimes
into a new spack.externals_config module.

The new module is intentionally free of spack.solver imports, which
allows it to be used by both the solver and other utilities
without creating circular import cycles.
